feat: move analytics charts and lifetime stats from History to Dashboard
Implements stage 1 of #170. MPG/price trend chart, cost-per-litre
chart, multi-vehicle comparison chart, and lifetime stat cards now
live on the Dashboard tab. History keeps the filterable summary
metrics, table/card log view, edit/delete, and export tools.
